% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_count_matrix.R
\name{extract_count_matrix}
\alias{extract_count_matrix}
\title{Extract ordered matrix from prepared heatmaps list object}
\usage{
extract_count_matrix(prepared_hm_list)
}
\arguments{
\item{prepared_hm_list}{A \code{HeatmapList-class} object for which the 
layout has been created and which has been drawn up in advance.}
}
\value{
A list of dataframes
}
\description{
Extract ordered matrix from prepared heatmaps list object
}
\examples{
\dontrun{
htl <- get_demo_heatmap_list()
htl <- prepare_heatmap_list(htl)
htl_df <- extract_count_matrix(htl)
}

}
