% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_bdg}
\alias{get_demo_bdg}
\title{Get a bedGraph file for demo}
\usage{
get_demo_bdg()
}
\value{
The path to a valid bdg file
}
\description{
Get a bedGraph file for demo
}
\examples{
bdg <- get_demo_bdg()

}
