% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_bdg_idx}
\alias{get_demo_bdg_idx}
\title{Get an indexed bedGraph file for demo}
\usage{
get_demo_bdg_idx()
}
\value{
The path to a valid indexed bdg file
}
\description{
Get an indexed bedGraph file for demo
}
\examples{
bdg <- get_demo_bdg_idx()

}
