% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_bed_files}
\alias{get_demo_bed_files}
\title{Get a bed file for demo}
\usage{
get_demo_bed_files()
}
\value{
The path to a valid bed file
}
\description{
Get a bed file for demo
}
\examples{
bed <- get_demo_bed_files()

}
