% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_heatmap_list}
\alias{get_demo_heatmap_list}
\title{Return a list of heatmaps for demo}
\usage{
get_demo_heatmap_list(partitions = FALSE)
}
\arguments{
\item{partitions}{Should the heatmaps be splitted? \code{TRUE} or
\code{FALSE}. Default: \code{FALSE}.}
}
\value{
A \code{list} of \code{Heatmap} objects
}
\description{
Return a list of heatmaps for demo
}
\examples{
bdg <- get_demo_heatmap_list()

}
