% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bedgraph.R
\name{import_bedgraph}
\alias{import_bedgraph}
\title{Import bedgraph files}
\usage{
import_bedgraph(filename, filter_negative_coverages = TRUE,
  keep_standard_chromosomes = TRUE, genome = NULL)
}
\arguments{
\item{filename}{Paths to a valid bedGraph file.}

\item{filter_negative_coverages}{Convert negative coverage values to 0?
(Default: TRUE)}

\item{keep_standard_chromosomes}{Remove alternative chromosomes? 
(Default: TRUE)}

\item{genome}{The name of the genome (i.e.: "hg38", "mm10", etc...)}
}
\value{
A list of GRanges (one element per file).
}
\description{
Import bedgraph files
}
\examples{
filenames <- get_demo_bdg()
cov <- import_bedgraph(filenames)

}
