% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_heatmap_list.R
\name{prepare_heatmap_list}
\alias{prepare_heatmap_list}
\title{Combine the heatmaps objects}
\usage{
prepare_heatmap_list(heatmap_list, partitions = character(0), seed = 99841)
}
\arguments{
\item{heatmap_list}{The \code{list} of \code{Heatmap} objects produced with
the \code{produce_heatmap} function.}

\item{partitions}{If you want to split the heatmap based on different group,
you must specify the groups (or partitions). The \code{partition} must be a
\code{vector} of the same length as the \code{peaks} param used when
creating the \code{Heatmap} objects with the \code{produce_heatmap}
function. All heatmaps must be produced using the same \code{peaks} object.
If the \code{vector} is empty (i.e.: \code{character(0)}, the heatmap won't
be splitted. Default: \code{character(0)}.}
}
\value{
This function invisibly returns a \code{HeatmapList-class} object
for which the layout has been created.
}
\description{
This function prepare the ready for drawing \code{HeatmapList-class} object.
}
\examples{
\dontrun{
htl <- get_demo_heatmap_list()
htl <- prepare_heatmap_list(htl)
}

}
