% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_heatmap.R
\name{produce_heatmap}
\alias{produce_heatmap}
\title{Convert coverage to heatmap}
\usage{
produce_heatmap(cov, peaks, column_title, force_seqlevels = FALSE)
}
\arguments{
\item{cov}{A coverage generated with the \code{import_bedgraphs} function}

\item{peaks}{A \code{GRanges} object representing the peaks to import}

\item{column_title}{The name of the heatmap that will be shown at the top}

\item{force_seqlevels}{If \code{TRUE}, remove regions that are not found
in the coverage. Corresponds to \code{pruning.mode =
"coarse"} in \code{?seqinfo}. Default: \code{FALSE}.}
}
\value{
An heatmap object
}
\description{
Convert coverage to heatmap
}
\examples{
bdg <- get_demo_bdg()
narrow_peak_file <- peakimport::get_demo_file("narrowPeak")
peaks <- peakimport::import_peaks(narrow_peak_file)
cov <- import_bedgraphs(bdg)
heatmap <- produce_heatmap(cov, peaks, "demo")

}
