% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_heatmap.R
\name{produce_heatmap}
\alias{produce_heatmap}
\title{Convert coverage to heatmap}
\usage{
produce_heatmap(cov, peaks, name, partitions = character(0),
  col = NULL, force_seqlevels = FALSE, seed = 99841,
  top_anno_ylim = NULL)
}
\arguments{
\item{cov}{A coverage generated with the \code{import_bedgraph} function.}

\item{peaks}{A \code{GRanges} object representing the peaks to import. Peaks
will be resized to 5000 base pairs to make sure all regions are of the same
length.}

\item{name}{The name of the heatmap that will be shown at the top.}

\item{partitions}{If you want to split the heatmap based on different group,
you must specify the groups (or partitions). The \code{partition} must be a
\code{vector} of \code{character} of the same length as the \code{peaks}
param. If the \code{vector} is empty (i.e. \code{character(0)}, the heatmap
won't be splitted. Default: \code{character(0)}.}

\item{col}{A \code{circlize::colorRamp2} object. If \code{NULL}, default
values will be used.}

\item{force_seqlevels}{If \code{TRUE}, remove regions that are not found
in the coverage. Corresponds to \code{pruning.mode =
"coarse"} in \code{?seqinfo}. Default: \code{FALSE}.}

\item{seed}{Set a seed value for the color selection. Only used when
\code{partition} is not \code{NULL}. The value must be the same as the one
used for the \code{prepare_heatmap_list} function. Default: 99841.}

\item{top_anno_ylim}{Manually set the ylim value for the top_annotation.
If \code{NULL}, value is inferred using the data. Default: \code{NULL}.}
}
\value{
An heatmap object
}
\description{
Convert coverage to heatmap
}
\examples{
bdg <- get_demo_bdg()
bed_file <- get_demo_bed_files()[[1]]
peaks <- peakimport::import_peaks(bed_file)
cov <- import_bedgraph(bdg)
heatmap <- produce_heatmap(cov, peaks, "demo")

}
