% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_heatmap.R
\name{produce_heatmap}
\alias{produce_heatmap}
\title{Convert coverage to heatmap}
\usage{
produce_heatmap(cov, peaks, name, force_seqlevels = FALSE)
}
\arguments{
\item{cov}{A coverage generated with the \code{import_bedgraphs} function.}

\item{peaks}{A \code{GRanges} object representing the peaks to import. Peaks
will be resized to 5000 base pairs to make sure all regions are of the same
length.}

\item{name}{The name of the heatmap that will be shown at the top.}

\item{force_seqlevels}{If \code{TRUE}, remove regions that are not found
in the coverage. Corresponds to \code{pruning.mode =
"coarse"} in \code{?seqinfo}. Default: \code{FALSE}.}
}
\value{
An heatmap object
}
\description{
Convert coverage to heatmap
}
\examples{
bdg <- get_demo_bdg()
bed_file <- get_demo_bed_files()[[1]]
peaks <- peakimport::import_peaks(bed_file)
cov <- import_bedgraphs(bdg)
heatmap <- produce_heatmap(cov, peaks, "demo")

}
